Move language related logic to language model file
